Move infer_rt to Separate Submodule
